requirements.txt
streamlit
yfinance
pandas
scikit-learn
requests
beautifulsoup4
newspaper3k
transformers
torch
python-dotenv 